Use Vite to serve and build server code. Fixes build.
