+++ .env.server
@@ -1,0 +1,1 @@
+NODE_ENV=production
