+++ .gitignore
@@ -9,8 +9,7 @@
 /coverage
 
 # production
-/build
-/server.bundle.js
+/dist
 
 # misc
 .DS_Store
